Create new wrapper for ExecutionCreateApi (#48)
The concept of repostitories/spaces etc. was leaking out of the Java SDK and into the client app - in that the client app became aware of the need to traverse into the space prior to performing the operation.

New wrappers have been put around the exeuctions api whch hides this detail, which will allow clients to be simpler.